+++ pysys/process/user.py
@@ -1967,7 +1967,7 @@ class ProcessUser(object):
 
 
 	def logFileContents(self, path, includes=None, excludes=None, maxLines=20, tail=False, encoding=None, 
-			logFunction=None, reFlags=0, stripWhitespace=True, mappers=[], color=True):
+			logFunction=None, reFlags=0, stripWhitespace=True, mappers=[], color=True, message=None):
 		""" Logs some or all of the lines from the specified file.
 		
 		If the file does not exist or cannot be opened, does nothing. The method is useful for providing key
@@ -2021,6 +2021,10 @@ class ProcessUser(object):
 		
 		:param bool color: By default logged lines are colored blue to distinguish from the rest of the log contents. 
 			Set this to False to disable coloring. Added in PySys 2.1. 
+		
+		:param str message: The introductory message to log before the file content, with ``@PATH@`` as a placeholder for the path. 
+			If not specified the default is equivalent to ``"Contents of @PATH@: "``. 
+			Added in PySys 2.2
 
 		:return: True if anything was logged, False if not.
 		
@@ -2080,7 +2084,9 @@ class ProcessUser(object):
 
 		path = os.path.normpath(path)
 		if path.startswith(self.output): path = path[len(self.output)+1:]
-		self.log.info(u'Contents of %s%s: ', fromLongPathSafe(path), ' (filtered)' if includes or excludes else '', 
+		self.log.info('%s',
+			message.replace('@PATH@', fromLongPathSafe(path)) if message else
+			u'Contents of %s%s: '%(fromLongPathSafe(path), ' (filtered)' if includes or excludes else ''), 
 			extra=BaseLogFormatter.tag(LOG_FILE_CONTENTS, suppress_prefix=False))
 		for l in tolog:
 			logFunction(l)